Change test case chrono now timestamp to u64

+++ src/quote/mod.rs
@@ -1198,7 +1198,7 @@ mod tests {
 
         let enclave_identity_root: EnclaveIdentityRoot = serde_json::from_str(include_str!("../../data/qeidentity.json")).unwrap();
 
-        let current_time = chrono::Utc::now().timestamp();
+        let current_time = chrono::Utc::now().timestamp() as u64;
 
         let json_str = include_str!("../../data/tcbinfo.json");
         let tcb_info_root: TcbInfoRoot = serde_json::from_str(json_str).unwrap();
